+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+* [`useForm`, `useFields`] Resetting fields now sets the value to empty string if no default value
+
 ### Changed
 * [`useConstraint`] Update the number regex pattern